pyproject.toml
[tool.poetry]
name = "balhom-djangorest-api"
version = "2.0.0"
description = ""
authors = ["fabbo-repo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.8"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
django-stubs = "^4.2.7"
django-configurations = "^2.5"
django-filter = "^23.5"
django-cors-headers = "^4.3.1"
drf-yasg = "^1.21.7"
pyyaml = "^6.0.1"
coreapi = "^2.3.3"
python-keycloak = "^3.7.0"
pyjwt = "^2.8.0"
pillow = "^10.1.0"
piexif = "^1.1.3"
celery = "^5.3.6"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.5.0"
redis = "^5.0.1"
eventlet = "^0.34.1"
uritemplate = "^4.1.1"
psycopg2-binary = "^2.9.9"
coverage = "^7.3.4"
beautifulsoup4 = "^4.12.2"
gunicorn = "^21.2.0"
django-storages = "^1.14.2"
boto3 = "^1.34.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
